refactor: move concrete policy definitions to model directory
