+++ vmngclient/endpoints/configuration/policy/definition_builder/data.py
@@ -1,75 +1,15 @@
 # mypy: disable-error-code="empty-body"
-from typing import List, Optional, Union
-
-from pydantic import Field
-from typing_extensions import Annotated
+from typing import Optional
 
 from vmngclient.endpoints import APIEndpoints, delete, get, post, put
+from vmngclient.model.policy.definitions.data import DataPolicy, DataPolicyDefinition
 from vmngclient.model.policy.policy_definition import (
-    AppListEntry,
-    DefinitionSequence,
-    DestinationDataIPv6PrefixListEntry,
-    DestinationDataPrefixListEntry,
-    DestinationIPEntry,
-    DestinationPortEntry,
-    DestinationRegionEntry,
-    DNSAppListEntry,
-    DNSEntry,
-    DSCPEntry,
-    PacketLengthEntry,
-    PLPEntry,
-    PolicyDefinitionBody,
     PolicyDefinitionEditResponse,
-    PolicyDefinitionHeader,
     PolicyDefinitionId,
     PolicyDefinitionInfo,
     PolicyDefinitionPreview,
-    ProtocolEntry,
-    SourceDataIPv6PrefixListEntry,
-    SourceDataPrefixListEntry,
-    SourceIPEntry,
-    SourcePortEntry,
-    TCPEntry,
-    TrafficToEntry,
 )
 from vmngclient.typed_list import DataSequence
-
-DataPolicySequenceEntry = Annotated[
-    Union[
-        PacketLengthEntry,
-        PLPEntry,
-        ProtocolEntry,
-        DSCPEntry,
-        SourceIPEntry,
-        SourcePortEntry,
-        DestinationIPEntry,
-        DestinationPortEntry,
-        TCPEntry,
-        DNSEntry,
-        TrafficToEntry,
-        SourceDataPrefixListEntry,
-        DestinationDataPrefixListEntry,
-        SourceDataIPv6PrefixListEntry,
-        DestinationDataIPv6PrefixListEntry,
-        DestinationRegionEntry,
-        DNSAppListEntry,
-        AppListEntry,
-    ],
-    Field(discriminator="field"),
-]
-
-
-class DataPolicySequence(DefinitionSequence):
-    type: str = Field(default="data", const=True)
-    entries: List[DataPolicySequenceEntry]
-
-
-class DataPolicy(PolicyDefinitionHeader):
-    type: str = Field(default="data", const=True)
-
-
-class DataPolicyDefinition(PolicyDefinitionBody):
-    sequences: List[DataPolicySequence] = []
 
 
 class DataPolicyCreationPayload(DataPolicy):
